Basic implementation
I should have started committing much, much earlier in this process,
but time flies when you are learning Rust.

That means you get to miss out on the initial, really, really badly
written version. I won't claim this is good or idiomatic Rust, but at
least I'm not currently ashamed of it.