Prune guess suggestions before doing more expensive computations
Computing all average remaining words counts is O(n^3) in the number
of remaining possibilities. I have some ideas for scores that are even
more algorithmically expensive. Rather than waste time computing these
scores for garbage guesses, we now drop the majority of guesses before
moving on to computing the more expensive scores. Theoretically, this
could cause it to miss an unusual suggestion, but in practice it seems
to be producing the exact same results a lot more quickly.