Show a warning early on if the answer isn't known to us
Thus far we are only aware of NYT removing words from the answer list,
but conceivable, they may someday start adding them. If that happens,
it'd be best to let the user know as soon as possible, because later
analysis steps could run into trouble.

Of course, there's also the perhaps more likely scenario of the user
just mistyping the answer.