Introduce a weighted green/yellow count and use it to order guesses
I wasn't happy with the old behaviour of comparing on the average
green/yellow count and only considering the average green count in the
event of a tie. It just didn't feel like the right way to do things.

Due to its semi-arbitrary nature, I don't want to really want to
expose this number in the UI, so I'll continue to compute and show the
old values while internally sorting by the new combined value.

The actual weightings are pulled from my butt, but they feels pretty
good. They are trivial to tweak in the future, should we want to.