Switch to clap for command line argument parsing
My first impressions of clap are that it feels to be both overkill and
a little too restrictive to me, but it seems to be the defacto
stardard in Rust and my needs here are not so advanced as to justify
going off the beaten track.

Note that the --words flag must now be used when passing the game on
the command line.