Add a --columns flag to control the width of the output
Currently, when running as a CGI script, the width of the output in
the webpage is entirely dependent on the width of the terminal from
which the webserver was launched. This is obviously inconvenient and
silly.

I had originally attempted to let the caller specify the width of the
output in characters, but that proved to be harder to implement and
more for the user to think about, so I just went with columns which
feels much simpler for everyone.