Optimise WorkKnowledge::matches by only checking for yellows
Previously this code was checking that the PossibleAnswer contains all
of the known letters. This was wasteful as checking that the slots
match already takes care of the green letters, so we only need to
worry about the yellow letters in the second pass.

I was shocked to find that this change alone yielded about a 19%
performance improvement when running --suggest-first-guess.