Rewrite Alphagram::contains_other for readability
I was super curious whether this new "higher level" version would be
faster or slower than the old one. The algorithm is slightly smarter in
a way that could make it faster, but it is also using fancy iterators
that could potentially obscure the compilers view of what's going on.

In the end, I'm not able to notice a difference in performance between
the two. Since we know this function is on our hot path, I'm going to
trust the language and the compiler and go with the prettier
implementation.