Introduce FastWordKnowledge
This is just a stripped down copy of WordKnowledge focused on speed.

WorkKnowledge::add_guess is some of my favourite code in this whole
program, so it hurts me to copy-and-paste and then hack-and-slash it
down like this. But the fact of the matter is that it handles a lot of
tricky scenarios we just don't need to worry about when assessing
potentially guesses.

When assessing player guesses, we build upon all past guesses to be
able to show a solve state containing all the information we have so
far. We also have to handle the player making errors or intentionally
conflicting guesses. Neither of these apply when doing the millions of
checks needed to assess potential guesses. We are only considering
remaining possibilities and we have already used past guesses to
filter out any possibilities they didn't support, so we don't need to
consider those past guesses any more. And we are never going to
suggest a conflicting guess, so we don't need to handle those
scenarios either.

Faster is faster.