Merge FastWordKnowledge back into WordKnowledge
FastWordKnowledge was a dumb idea.

My IDE kindly pointed out that FastWordKnowledge::matches was an exact
duplicate of WordKnowledge::matches, which was enough of a slap to the
head to make me rethink why I had created FastWordKnowledge to begin
with. My two motivators were fast construction from a guess and
avoiding the need for an all-letter histogram. Neither of these
required creating a new struct.