Promote Letter to a real struct from a u8 alias
This is a big change. The diff tells me that it is adding ~40 lines of
code, but it actually feels like a pretty big simplification.

This was initially done in an attempt to optimise LetterSet
performance, which it does by a small but measurable amount, but it
quickly spread into a bunch of other improvements as well. I probably
could have done better at limiting the scope of this commit, but one
thing lead to the next and hear we are.

* We now use NO_LETTER everywhere rather than a bunch of different
placeholder letter values.
* A lot more iteration of Letters is by value, which generally feels
nicer. Fewer ampersands, at least.
* We no longer do any ASCII uppercase conversions or checks.
* Letter::is_valid makes a bunch of lines just slightly nicer.