Save one pass over the possibilities in best_guesses()
This change has been living in a stash for a long time. Not sure why I
never applied it.

I played with using collect_into_vec as I suspected that might be
faster, but I didn't see any difference in performance between regular
collect() and collect_into_vec() with a preallocated Vec.