Implement Display for Word
My printing and formatting functions were very much written by a Rust
first-timer and it shows. I really liked the simplicity of the colored
crate when first starting out, but now I'm not so sure it's the best
choice.

These functions are still crap, but I think these changes are a step
forward. Fewer string allocations are always nice, evwen if I don't
particularly care about the performance of the output code.

Presumably, the goal should be to separate the output code from the
rest of the logic so that I could, say, natively generate HTML _or_
colored terminal output.