Add five new words to the answer list
We now know that the NYT's Wordle editor is actively editing the word
list. As the answer list is no longer public, we can't know which
words have been added or removed, except when a new word appears as
the answer to a daily puzzle. Unbeknownst to me, that happened 4 times
in the spring of last year (BALSA, GUANO, KAZOO, SNAFU), but it
happened again this week (LASER).

I was a little bit torn on whether to add the words or not, since they
are extremely unlikely to be used as answers again anytime soon, but
as I don't remove words from the list as they are used, it would seem
inconsistent to not include them.