Make suggestion order more deterministic
In the case of the game ARISE -> DIARY, Tortordle was suggesting RADIO
as its first suggested guess, but then chose RADII when attempting to
solve by itself.

That fact was that the two words had identical scores, but because our
sorting of suggestions wasn't completely deterministic, the fact that
attempt_optimal_solve only ever asked for the one top guess meant that
it ended up getting a different suggestion than if you asked for the
12 top guesses and took the first one. Not strictly a bug, but also a
little surprising or confusing to users.

Making the ordering entirely deterministic doesn't cost us much and
just feels right.