Significantly simplify Alphagram merging
* Renamed the function to merged().
* Made it return the merged result rather than modify the first
parameter.
* Upgraded index-based while loop to a simple for loop.
* Removed handling of two different scenarios that can't ever happen.
* Added debug_assert!s to check that the merged result always fits in
the result. This should always be the case in our use of the class.

It feels like I should be able to simplify this further by replacing i
and j with iterators, but I don't actually think that's the case. I
can make it work with iterators, but the result is never simpler.