Introduce ColorCounts
Tortordle contains two very hot loops that create a great many
ColoredGuesses only to loop over them to total up the number of greens
and yellows they contain. The compiler wasn't able to completely
optimize away these temporary ColoredGuesses, leaving room for a hand
optimization and simplification of the calling code. Of course this
comes at the cost of a second version of the guess colouring function.