Drop the macro from sorted5
Rust 1.83 now allows mutable references in const references, which
lets us get away with a simple function instead of a macro.

Sadly, std::mem::swap didn't get stabilized in const contexts in time
for this release, so we still need to do the temp dance. Hopefully
that can go with Rust 1.84.